bignum: force size of vm vaargs integer arguments

+++ src/kex_dh.c
@@ -127,7 +127,7 @@ static assh_error_t assh_kex_dh_client_send_expmod(struct assh_session_s *s)
 
   ASSH_JMP_ON_ERR(assh_bignum_bytecode(c, 0, bytecode, "MMMNNsTTm",
                    /* M */ gr->generator, gr->prime, e_str,
-                   /* N */ &pv->en, &pv->xn, gr->size), err_p);
+                   /* N */ &pv->en, &pv->xn, (size_t)gr->size), err_p);
 
   assert(pv->xn.secret);
 
@@ -216,7 +216,7 @@ static ASSH_EVENT_DONE_FCN(assh_kex_dh_host_key_lookup_done)
 
   ASSH_JMP_ON_ERR(assh_bignum_bytecode(s->ctx, 0, bytecode, "MMMMNsTTTTTm",
                    /* M */ gr->generator, gr->prime, f_str, secret,
-                   /* N */ &pv->xn, gr->size)
+                   /* N */ &pv->xn, (size_t)gr->size)
 	       | ASSH_ERRSV_DISCONNECT, err_scratch);
 
   ASSH_JMP_ON_ERR(assh_hash_init(s->ctx, hash_ctx, gr->hash)
@@ -364,7 +364,7 @@ static assh_error_t assh_kex_dh_server_wait_e(struct assh_session_s *s,
 
   ASSH_JMP_ON_ERR(assh_bignum_bytecode(c, 0, bytecode, "MMMMMssTTTTTTTm",
                    gr->generator, gr->prime, e_str, f_str, secret,
-                   pv->exp_n, gr->size), err_p);
+                   (size_t)pv->exp_n, (size_t)gr->size), err_p);
 
   assh_packet_string_resized(pout, f_str + 4);
